Move things around to better establish legacy vs non-legacy
